fix: fix the project so that the first migration can be done

+++ src/Backend/CozySoccerChamp.Domain/Entities/Soccer/Prediction.cs
@@ -8,7 +8,7 @@ public class Prediction : BaseEntity
     public int MatchId { get; set; }
     public int? PredictedHomeScore { get; set; }
     public int? PredictedAwayScore { get; set; }
-    public int? PointPerMatch { get; set; }
+    public double? PointPerMatch { get; set; }
     public DateTime PredictionTime { get; set; }
 
     public virtual ApplicationUser User { get; set; }
